Pull in latest dev branch changes and harmonize with new modules
