+++ tools/RAiDER/geometry.py
@@ -106,49 +106,6 @@ def checkLOS(los, Npts):
 # Part of the following UTM and WGS84 converter is borrowed from https://gist.github.com/twpayne/4409500
 # Credits go to Tom Payne
 
-_projections = {}
-
-
-def zone(coordinates):
-    if 56 <= coordinates[1] < 64 and 3 <= coordinates[0] < 12:
-        return 32
-    if 72 <= coordinates[1] < 84 and 0 <= coordinates[0] < 42:
-        if coordinates[0] < 9:
-            return 31
-        elif coordinates[0] < 21:
-            return 33
-        elif coordinates[0] < 33:
-            return 35
-        return 37
-    return int((coordinates[0] + 180) / 6) + 1
-
-
-def letter(coordinates):
-    return 'CDEFGHJKLMNPQRSTUVWXX'[int((coordinates[1] + 80) / 8)]
-
-
-def project(coordinates, z=None, l=None):
-    if z is None:
-        z = zone(coordinates)
-    if l is None:
-        l = letter(coordinates)
-    if z not in _projections:
-        _projections[z] = pyproj.Proj(proj='utm', zone=z, ellps='WGS84')
-    x, y = _projections[z](coordinates[0], coordinates[1])
-    if y < 0:
-        y += 10000000
-    return z, l, x, y
-
-
-def unproject(z, l, x, y):
-    if z not in _projections:
-        _projections[z] = pyproj.Proj(proj='utm', zone=z, ellps='WGS84')
-    if l < 'N':
-        y -= 10000000
-    lng, lat = _projections[z](x, y, inverse=True)
-    return (lng, lat)
-
-
 def WGS84_to_UTM(lon, lat, common_center=False):
     shp = lat.shape
     lon = np.ravel(lon)
@@ -172,7 +129,12 @@ def WGS84_to_UTM(lon, lat, common_center=False):
         L[ind] = l
         X[ind] = x
         Y[ind] = y
-    return np.reshape(Z,shp), np.reshape(L,shp), np.reshape(X,shp), np.reshape(Y,shp)
+    return (
+            np.reshape(Z,shp), 
+            np.reshape(L,shp), 
+            np.reshape(X,shp), 
+            np.reshape(Y,shp)
+        )
 
 def UTM_to_WGS84(z, l, x, y):
     shp = x.shape
@@ -191,3 +153,43 @@ def UTM_to_WGS84(z, l, x, y):
         lat[ind] = coordinates[1]
         lon[ind] = coordinates[0]
     return np.reshape(lon,shp), np.reshape(lat,shp)
+
+
+def project(coordinates, z=None, l=None):
+    if z is None:
+        z = zone(coordinates)
+    if l is None:
+        l = letter(coordinates)
+    _proj = pyproj.Proj(proj='utm', zone=z, ellps='WGS84')
+    x, y = _proj(coordinates[0], coordinates[1])
+    if y < 0:
+        y += 10000000
+    return z, l, x, y
+
+
+def unproject(z, l, x, y):
+    _proj = pyproj.Proj(proj='utm', zone=z, ellps='WGS84')
+    if l < 'N':
+        y -= 10000000
+    lng, lat = _proj(x, y, inverse=True)
+    return (lng, lat)
+
+
+def letter(coordinates):
+    return 'CDEFGHJKLMNPQRSTUVWXX'[int((coordinates[1] + 80) / 8)]
+
+
+def zone(coordinates):
+    if 56 <= coordinates[1] < 64 and 3 <= coordinates[0] < 12:
+        return 32
+    if 72 <= coordinates[1] < 84 and 0 <= coordinates[0] < 42:
+        if coordinates[0] < 9:
+            return 31
+        elif coordinates[0] < 21:
+            return 33
+        elif coordinates[0] < 33:
+            return 35
+        return 37
+    return int((coordinates[0] + 180) / 6) + 1
+
+
